Refactoring
- Removed useless sort modes
- Removed unused classes
- Refactored Trip Store and Quote Store
- Refactored packages of multiple classes
