+++ models/HTTP/Header.java
@@ -1,4 +1,4 @@
-package main.java.models.HTTP;
+package main.java.models.http;
 
 /**
  * A class that represents a header used in HTTP requests.
