--- models/Trips/enums/TripSortMode.java
@@ -1,11 +1,0 @@
-package main.java.models.Trips.enums;
-
-/**
- * A simple ENUM that defines Trip Sort Modes
- *
- * @author Teegan Krieger
- */
-public enum TripSortMode {
-    CHEAPEST,
-    EXPENSIVE
-}
