+++ models/flightapi/collections/RoutesStore.java
@@ -1,4 +1,4 @@
-package main.java.models.flightapi.stores;
+package main.java.models.flightapi.collections;
 
 /**
  *
