--- models/flightapi/comparators/QuoteAlphabeticalComparator.java
@@ -1,22 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package main.java.models.flightapi.comparators;
-
-import java.util.Comparator;
-import main.java.models.flightapi.structures.UniversalQuote;
-
-/**
- *
- * @author teega
- */
-public class QuoteAlphabeticalComparator implements Comparator<UniversalQuote> {
-
-    @Override
-    public int compare(UniversalQuote o1, UniversalQuote o2) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-}
