--- models/flightapi/structures/Route.java
@@ -1,9 +1,0 @@
-package main.java.models.flightapi.structures;
-
-/**
- *
- * @author STUB
- */
-public class Route {
-    
-}
