+++ models/trips/collections/TripsStore.java
@@ -1,55 +1,55 @@
-package main.java.models.Trips.stores;
+package main.java.models.trips.collections;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import main.java.models.Trips.comparators.TripCheapestComparator;
-import main.java.models.Trips.comparators.TripExpensiveComparator;
-import main.java.models.Trips.Trip;
-import main.java.models.Trips.enums.TripSortMode;
+import main.java.models.trips.comparators.TripCheapestComparator;
+import main.java.models.trips.comparators.TripExpensiveComparator;
+import main.java.models.enums.SortMode;
+import main.java.models.interfaces.IStore;
 
 /**
  * A class that stores loaded trips and allows for sorting of these trips
  *
  * @author Teegan Krieger
  */
-public class TripsStore {
+public class TripsStore implements IStore {
 
     private static final TripCheapestComparator tripCheapestComparator = new TripCheapestComparator();
     private static final TripExpensiveComparator tripExpensiveComparator = new TripExpensiveComparator();
 
     private ArrayList<Trip> sortedTrips;
 
-    private TripSortMode currentSortMode;
+    private SortMode currentSortMode;
 
     public TripsStore() {
         sortedTrips = new ArrayList<>();
-        currentSortMode = TripSortMode.CHEAPEST;
+        currentSortMode = SortMode.CHEAPEST;
     }
 
     /**
      * Adds 1 or more Trips to the Store
      * @param _objects
      */
-    public void addCollection(Trip... _objects) {
+    public void addCollection(Object... _objects) {
 
         for (int i = 0; i < _objects.length; i++) {
-            sortedTrips.add(_objects[i]);
+            sortedTrips.add((Trip)_objects[i]);
         }
 
         sort(currentSortMode);
     }
-    
+
      /**
      * Adds 1 Trips to the Store and Give it a free Name by appending a number if needed.
      * @param _t the tip.
-     * @pram _tripName the name to use. 
+     * @pram _tripName the name to use.
      */
     public void addCollection(Trip _t, String _tripName) {
-        
+
         _t.setName( this.getNextFreeName(_tripName));
-        
+
         this.sortedTrips.add(_t);
-        
+
 
         this.sort(this.currentSortMode);
     }
@@ -73,7 +73,7 @@ public class TripsStore {
      * Sort the Trips within the store using the given sort mode
      * @param _sortMode The sort mode
      */
-    public void sort(TripSortMode _sortMode) {
+    public void sort(SortMode _sortMode) {
         switch (_sortMode) {
             case CHEAPEST:
                 Collections.sort(sortedTrips, tripCheapestComparator);
@@ -111,7 +111,7 @@ public class TripsStore {
 
         return allTrips;
     }
-    
+
     /**
      * get a string that is not currently a tripName by appending a number or Incrementing a number at the end.
      * @param _tripName The new name for the trip.
@@ -119,30 +119,30 @@ public class TripsStore {
      */
     public String getNextFreeName( String _tripName)
     {
-         ArrayList<String> namesList = new ArrayList<String>();        
-        
+         ArrayList<String> namesList = new ArrayList<String>();
+
         if(_tripName.trim().equals(""))
             _tripName = "New Trip";
-         
+
         for (Trip t : this.sortedTrips)
         {
             namesList.add(t.getName());
         }
-        
+
         while (namesList.contains(_tripName))
         {
             if (_tripName.contains("-"))
             {
                 String[] a = _tripName.split("-", -2);
-            
+
                 int newNumber = Integer.parseInt(a[1]) + 1;
                 _tripName = a[0] + "-" + newNumber;
             }
             else
                 _tripName += "-1";
-            
+
         }
-        
+
         return _tripName;
     }
